Zealot Reorg
The zealot folder is now mainly a node typescript project.
npm run build will create a bundled commonjs and es module version
of the package. It will also generate a index.d.ts with all the
types to carry the exported types into another typescript project.

The test_api folder is a deno project and imports the bundled es
module package to use and run tests agains a real zqd.

The zjson file contains types for the new zjson format

The zqd folder contains types for the expected payloads.

The zng folder contains the classes that wrap the zjson data.