zq update through "Convert eve.json upon ingest" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1400, authored by @henridf,
has been merged.

Convert eve.json upon ingest

This PR adds a json type config that is suited for "alert" events output by the Brim suricata package. The config is used when transforming the Suricata-produced "eve.json" file. In addition, the suricata `timestamp` field is renamed to `ts`.

This PR is based on brimsec/zq#1389 so if you come here first, please try to make the detour.

Closes brimsec/zq#1213