Brim ZJSON Update (#1108)
* Zealot Reorg

The zealot folder is now mainly a node typescript project.
npm run build will create a bundled commonjs and es module version
of the package. It will also generate a index.d.ts with all the
types to carry the exported types into another typescript project.

The test_api folder is a deno project and imports the bundled es
module package to use and run tests agains a real zqd.

The zjson file contains types for the new zjson format

The zqd folder contains types for the expected payloads.

The zng folder contains the classes that wrap the zjson data.

* Auto Generate ZQD Test Respones

This script will run a zql search on a zqd instance that has
imported sample.tsv, a zeel file with 30 records. Then the output
of that search is piped to a src/js/test/responses folder that the
brim app uses to mock zqd in unit tests.

* Stop persisting log details

* Update Config Files

* Update Brim To Rely on ZNG Objects

The Log model and all usage has been removed
The brim/record model and all usage has been removed
The brim/field model has been changed to brim/cell and acts
like a decorator to a zng.Field class.

* Update Tests

* Pull in updated zq

* Update new error message assertion

* Pull in zq master