zq update through "archive.Import: implement buffered streaming" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1397, authored by @mattnibs,
has been merged.

archive.Import: implement buffered streaming

Previously archive.Import would read the entirety of the provided
stream, spilling to disk if necessary, before reading the stream
back in the order desired.

The change takes advantage of new overlapping chunk functionality.
Streams are now read into a zng buffer that is then sorted and written
to the archive once the buffer has:
1. Reached maxsize.
3. The provided zng stream has ended.

Part of brimsec/zq#1132