zq update through "add a type alias for port when reading azng" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1441, authored by @alfred-landrum,
has been merged.

add a type alias for port when reading azng

Recent Brim integration tests failures occurred because they were unable to read an alpha zng file that used the port type. This change automatically adds an alias for port -> uint16 whenever the azng reader is used, and adds a ztest with a more complex alpha zng file.

While testing, I discovered the alpha tzngio writer was using the post-alpha zcode package, so I pulled in the alpha zcode package as well.