zq update through "add Replace function to iosrc" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1447, authored by @alfred-landrum,
has been merged.

add Replace function to iosrc

This adds a Replace function to the iosrc package, with updates to the fs and s3 code to support it. I did this while working on the alpha zng migration in zqd.