zq update through "consolidate microindex Open into NewFinder" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1449, authored by @alfred-landrum,
has been merged.

consolidate microindex Open into NewFinder

This also fixes a panic that would be raised when an error occurred
opening the microindex that left some Reader fields nil, and uses the
correct error check for the "-Q" option to zar find.

Fixes brimsec/zq#1419