zq update through "Link fixes mostly related to ZNG spec update" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1457, authored by @philrz,
has been merged.

Link fixes mostly related to ZNG spec update

While it's convenient to be able to link directly to anchors for specific sections in docs as we've been doing, one of the hazards is that when a destination changes, the link doesn't show up as "broken" by our automation because the user just ends up being redirected to the top of the page in question. brimsec/zq#1394 renumbered some sections in the ZNG spec, so here I've manually rechecked relevant anchor links and made sure they go to the right places again.

While I was at it, I found some absolute URLs that I changed to be relative. This makes for better links when working in branches.