zq update through "separate compilation for driver.Run calls" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1452, authored by @alfred-landrum,
has been merged.

separate compilation for driver.Run calls

Instead of creating a fake multisource for calls to driver.Run, break
out the compilation for single reader vs multisource. This makes
upcoming changes to multisource easier to manage.