zq update through "change error to ErrNoSuchField accessing field of non-record" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1473, authored by @mccanne,
has been merged.

change error to ErrNoSuchField accessing field of non-record

This commit changes the dot accessor expression to return a
no-such-field error consistent with behavior prior to the
expr package overhaul.  This is important because group-by
gladly ignores keys that don't exist and should do so in the
case of accessing fields in data types that don't happen to
be a record.  This is a feature of the robust way zq/zng
handles queries over heterogeneously typed records.

Fixes brimsec/zq#1415