zq update through "add where clause to reducers" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1490, authored by @mccanne,
has been merged.

add where clause to reducers

This commit adds a "where" clause to reducers allowing you to
apply filtering to individual aggregations in a multi-column
group-by.  e.g.,

  lg=sum(foo) where x>100, sm=sum(foo) where x<100 by bar

This is needed in the beacon project.

This was also the right time to simplify the reducer syntax in the
peg grammar so that there is no longer an enumerated list of the available
aggregation functions.  Instead, the compiler pass makes sure the
the named reducer exists making it easier from now on to add reducers
without having to modify the grammar.