zq update through "add union aggregator" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1493, authored by @mccanne,
has been merged.

add union aggregator

This commit adds a union aggregator that takes union over all of its
input values and produces a value of type set of this computed union.
The type of the input values must be uniform.