zq update through "switch to ztest for reducer tests" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1508, authored by @mccanne,
has been merged.

switch to ztest for reducer tests

(This subsumes PR brimsec/zq#1503, which is closing without merge...)

This commit switches the reducer package from using reducer_test.go over
to ztest.

We designed the ztests so they would exercise the partials path and also
give empty partials to the partial consumer, which needs to be handled
correctly.

Along the way we fixed bugs in collect/union for handling empty partials
and noticed first/last should not be decomposable and thus should not
implement the reducer.Decomposable interface.