zq update through "Move link checker to separate Workflow" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1510, authored by @philrz,
has been merged.

Move link checker to separate Workflow

(This is equivalent to what just merged in https://github.com/brimsec/brim/pull/1153.)

The automation for the link checker has been doing its job insofar as it's noticed when external links go bad or act flaky. However, it's also been correctly noted that having it run as a Job in the same Action with the rest of ci.yml can be troublesome, since it makes it more difficult to debug when there's link problems at the same time as legit code breakages, or CI infra instability.

To address this, here we break out the link checker to its own separate Actions Workflow.