zq update through "merge spaninfos with the same largest chunk" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1486, authored by @alfred-landrum,
has been merged.

merge spaninfos with the same largest chunk

This reduces the number of SpanInfo's needed to read for a query when contiguous SpanInfo's share the same largest chunk, measured by record count.

An example: imagine chunk1's span is 0, 20, and it has 100 records; chunk2's span is 10, 20 and it has 10 records. Without this change, we would generate 3 SpanInfos: 0-10 (chunk1),10-20 (chunk1,chunk2), and 20-30 (chunk2). With this change, we'll generate 2 SpanInfo's: 0-20 (chunk1, chunk2), and 10-20 (chunk2).

This should help reduce the overall impact when a query includes a mix of large & small chunks, as might happen due to  longconn records.