zq update through "Replace output-format with output-flags in ztests" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1527, authored by @nwt,
has been merged.

Replace output-format with output-flags in ztests

The output-format key in ztest YAML files is redundant to output-flags.
Accordingly, replace "output-format: x" with "output-flags: -f x" in
YAML files and remove support for output-format from the ztest package.