zq update through "microindex: support descending indexes" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1526, authored by @mattnibs,
has been merged.

microindex: support descending indexes

Add support for microindexes created with a sorted flow of records
in descending order.

Additionally this pr adds ClosetGTE and ClosetLTE methods to
microindex.Finder: Functionality that will be needed for finding the
correct offsets for archive chunk seek indicies.

Closes brimsec/zq#1329