zq update through "move chunk metadata to separate file" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1528, authored by @alfred-landrum,
has been merged.

move chunk metadata to separate file

This PR moves the first & last timestamp, and the record count of a chunk, from the name of the chunk file to a separate metadata file. While working on compaction for overlapping files, I ran into local filesystem (not S3) limitations on the maximum length of a file name, which implied I'd need to store the masked chunk ids in a separate file. Since we plan on eventually storing metadata about a chunk to use for query optimization, I think it makes sense to encode all of the existing meta info to a separate file now, and add more content to it later as needed.

A repercussion of storing the timestamps this way is that all of the metadata files in a tsdir must be consulted at query time to determine which chunks to include for an query's time span. However, the metadata is immutable, so it can easily & safely be cached keyed on the chunk id.