zq update through "Fixes to pathnames changed in zq/1461" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1539, authored by @philrz,
has been merged.

Fixes to pathnames changed in zq/1461

The pathnames referenced in the [`zar` README](https://github.com/brimsec/zq/blob/master/cmd/zar/README.md) changed in brimsec/zq#1461, so this brings them current.

I also noticed the change to handle `port` as an alias had not yet been reflected in the example outputs, so I've addressed that as well.

Note that some file & byte counts shown in the README still do not match what you'll see if you run through the steps, due to brimsec/zq#1476.

Closes brimsec/zq#1530.