zq update through "ZQL docs for aggregate functions: and, or, collect, union" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1534, authored by @philrz,
has been merged.

ZQL docs for aggregate functions: and, or, collect, union

This PR adds ZQL docs for the recently-added aggregate functions `and()`, `or()`, `collect()`, and `union()`. I cleaned up a touch of other formatting while I was in there.

I've also noticed that the stuff under https://pkg.go.dev/github.com/brimsec/zq that we link to from the Developer Docs entries only seems to get updated when we tag a new release. Therefore the Developer Docs links I've added to these new reducers are not "broken" per se but they don't land at meaningful content yet. Likewise, some of the existing links still point at valid content that's been recently moved as part of code re-org and are going to need to be adjusted when we bundle the next release. Rather than trying to be fancy about timing the merge of docs PRs at the precise moment we bundle the next release, I'm inclined to merge docs now for the benefit of bleeding edge users working off the new stuff `master`. I've made a note-to-self to check/update links when I assemble the next GA release.