zq update through "support compaction of overlapping chunks" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1531, authored by @alfred-landrum,
has been merged.

support compaction of overlapping chunks

This change adds the `zar compact` command, with an optional `-purge` flag, that will look for runs of overlapping chunk files and write their data into a single chunk. As described in brimsec/zq#1183, the newly written chunks "mask" the existing chunks, so that at query time, if both an original chunk and a "compacted" chunk have data for the same time span, only the data from the compacted chunk will be used. This implementation records any masked chunk ids in a chunk's metadata file.

Closes brimsec/zq#1266.

(The high line count is due to the addition of test files.)