zq update through "fix bug where right bracket not properly escaped" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1561, authored by @mccanne,
has been merged.

fix bug where right bracket not properly escaped

This commit fixes a bug where the right bracket appearing as a
first character of a tzng value was not escaped in accordance
with the zng spec.

Fixes brimsec/zq#1305