zq update through "approximate compression size of tsDir record buffer" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1565, authored by @mattnibs,
has been merged.

approximate compression size of tsDir record buffer

For zar import, a tsDir was not taking account of compression size when
deciding if it had enough record data to write a chunk file. This results in
chunk files roughly half the size of log threshold size. Crudely approximate
size of uncompressed records by dividing buffer size by 2.

Closes brimsec/zq#1476