zq update through "move Cutter and ColumnBuilder out of proc" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1577, authored by @mccanne,
has been merged.

move Cutter and ColumnBuilder out of proc

This commit moves the Cutter and ColumnBuilder out of the proc
package so they can be reused by expr, which will be needed by
the cut() function.

ColumnBuilder was moved to its own package under zng/builder since
this purely a zng function to construct zng record values.  The Cutter
type has logic to take a stream of zng records and create "cuts" of
those value so this more complicated machinery was moved to package
expr.