zq update through "check and convert alpha zng filestores on access" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1574, authored by @alfred-landrum,
has been merged.

check and convert alpha zng filestores on access

When a filestore's data file is opened for queries, migrate it from alpha zng to beta zng if needed. This adds a field to the filestore's `info.json` file to record if the migration check is needed or not. A separate PR will add the ability to perform the check as a background routine in zqd, so that filestores can potentially be migrated before a user tries to access them.

As @philrz requested, I added log lines for migration start, success, and failure:
```
{"level":"info","ts":1604702663.6730611,"msg":"Alpha zng migration start","space_id":"sp_1jw5vTsJq5YfGZC5Z5Je2hT3J7A","file":"/Users/alfred/Library/Application Support/Brim/data/spaces/sp_1jw5vTsJq5YfGZC5Z5Je2hT3J7A/all.zng"}
{"level":"info","ts":1604702663.775091,"msg":"Alpha zng migration success","space_id":"sp_1jw5vTsJq5YfGZC5Z5Je2hT3J7A","file":"/Users/alfred/Library/Application Support/Brim/data/spaces/sp_1jw5vTsJq5YfGZC5Z5Je2hT3J7A/all.zng"}
```