zq update through "convert alpha zng spaces in background" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1576, authored by @alfred-landrum,
has been merged.

convert alpha zng spaces in background

When zqd starts, it will begin migrating any alpha zng filestores one at a time in the background. If a search request comes in for a space while its data is being migrated, the search will wait for the migration to complete.

Depends on https://github.com/brimsec/zq/pull/1574 .