zq update through "zar README updates due to zq/1565" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1571, authored by @philrz,
has been merged.

zar README updates due to zq/1565

As noted in brimsec/zq#1476, the compression changes in brimsec/zq#1565 changed some of the outputs shown in the `zar` README due to changes in file sizes/counters. I've re-run the README steps with the newest `zar` and updated those outputs here.