zq update through "refactor expr.Function" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1575, authored by @mccanne,
has been merged.

refactor expr.Function

This commit refactors expr.Function into a more modular approach,
changing the type from a function pointer to an interface whose
single method is Call().  This allows us to embed state in a function
that can be used to implement complex machinery, e.g., a cut()
function that behaves like a cut proc (caching lots of information
about types and builders) or a function to access an external
service like a key-value store (e.g., to do joins with thread intel
data feeds).

This change also allows us to clean up the temporary hack where
we passed the result buffer to a previously stateless function
in its Args parameters (Args is now replaced with []zng.Value).

The function names all now have a snake_case name in addition
to the package-style syntax.  The package syntax will be deprecated
in a subsequent commit.

We also changed the expression compiler to take a type context,
which will be used by functions and expression that need this,
which are coming soon.

Closes brimsec/zq#1524.