zq update through "add cut function" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1585, authored by @mccanne,
has been merged.

add cut function

This commit adds a cut function that behaves like the cut proc
but in the function context.  Like the proc, the cut function
ignores records that do not contain all the fields being cut.