zq update through "update coding style in peg grammar to conform with CamelCase " by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1587, authored by @mccanne,
has been merged.

update coding style in peg grammar to conform with CamelCase

This commit updates the coding style in the peg grammar to follow
the CamelCase convention.  Also, a few names were changed and
some rules were moved around to improve readability.  Some unused
rules were removed.  This is a style change only and does not include
any semantic changes.

This PR is arranged in three commits.  The first one is the name change,
the second one moves some peg rules around, and the third one is
the auto-generated peg parsers.