zq update through "Rename zbuf.RecordCmpFn to zbuf.RecordLessFn" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1591, authored by @nwt,
has been merged.

Rename zbuf.RecordCmpFn to zbuf.RecordLessFn

A zbuf.RecordCmpFn is a less function (like sort.Interface.Less) rather
than a comparision function (like bytes.Compare).  Rename it to
zbuf.RecordLessFn to reflect that, and rename functions and variables
accordingly.