zq update through "Advance Zeek pointer and fix tests" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1607, authored by @philrz,
has been merged.

Advance Zeek pointer and fix tests

We'd not advanced our Zeek pointer for the `zq` system tests in a while, so I've done that here.

Doing so revealed a couple tests that needed to be fixed. A connection that previously lacked a match in the geolocation database now has an entry, leading to different output. Much like we do for Zeek `uid` values that change on every run, here I've excluded the `geo` field so the tests won't be sensitive to this going forward. We already have decent upstream test coverage of the geolocation support in that the [smoketest Action in geoip-conn](https://github.com/brimsec/geoip-conn/blob/master/.github/workflows/smoketest.yml) ensures that each time we update the Zeek package it still outputs a location for a known location that we've never seen change/disappear.

This work is a prerequisite to a PR I'll put up next with a proposal to have a single set of tags for Zeek/Suricata dependencies in the zq repo and have those ripple to Brim when that gets built. I wanted to get everything current and passing before I go varying the implementation.