zq update through "deprecate package-style function names" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1609, authored by @mccanne,
has been merged.

deprecate package-style function names

This commit deprecates package-style function names by raising
an error with a helpful message for the alternative use.
We also eliminated some of the string functions in favor
of a type cast style, and extended the semantics of
numeric casts to accept strings (and parse the string
to a numeric value).  We also got rid of several of the
time conversion functions in favor of casts and adopted
the convention that a numeric value cast to time is in
units of nanosecond and added helper functions to convert
sec, msec, and usec to nanoseconds.

Closes brimsec/zq#1379