zq update through "rework merge/combiner" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1593, authored by @mccanne,
has been merged.

rework merge/combiner

This commit renames the orderedmerge proc as merge and
the old merge proc as combiner.  Toward this end, we
moved the merge logic into zbuf and changed the merge
proc to use it.  We also moved the scanner into zbuf.

This new arrangement allows Scanners to be created and
merged into a single puller without creating an
unnecessary proc object.  When we change the zar commands
to take pullers instead of readers (see brimsec/zq#1614), we can pass
a zbuf.Merger in as the puller and this will prevent an
extra copy of zng.Records to unbatch them through the
Read interface method.

The logic in zbuf.Merger could be improved if Batch.Records()
were changed to return a slice of zng.Record instead of
*zng.Record.  This is described in issue brimsec/zq#1613.