update brim to handle absent match-all filter at start of ast
A recent change to the peg parse modified ASTs so that they
do not have a "filter *" at the front of every AST, implying
the absence of a filter is the same as matching everything.
The fix here is to update the heuristic for finding a match-all
filter in an AST to include the case where there is no filter proc.