zq update through "add -P flag to zq" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1628, authored by @mccanne,
has been merged.

add -P flag to zq

This commit adds a new flag (-P) to zq to interconnect two or
more inputs directly to a zql query that begins with a parallel
flow graph.  This is needed by the forthcoming join proc.

Closes brimsec/zq#1616