zq update through "Fix type null handling in zng.Parse" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1631, authored by @nwt,
has been merged.

Fix type null handling in zng.Parse

zng.Parse translates an ast.Literal with type null to a zng.Value with a
Type field of nil instead of zng.TypeNull.  Fix that by allowing
zng.TypeNull.Parse to succeed if its parameter is empty.

Closes brimsec/zq#1387.