zq update through "fix bug with parents slice corruption in compiler" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1636, authored by @mccanne,
has been merged.

fix bug with parents slice corruption in compiler

This commit fixes a bug where the parents slice was overwritten
and reused so the parent linkage in the compiled flowgraph
becomes corrupt.  The fix is to make a new parents slice instead
of trying to reuse the buffer.

Fixes brimsec/zq#1635