zq update through "remove reminder comments in driver that were confirmed with henri" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1639, authored by @mccanne,
has been merged.

remove reminder comments in driver that were confirmed with henri

This commit removes XXX comments that were supposed to be removed
before the field data structure changes.  They were reminders to
check with Henri regarding the correctness of the query planner
transformations.  He has confirmed that the code is correct so
these comments can now be removed.