zq update through "fix join key expressions and add test coverage" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1642, authored by @mccanne,
has been merged.

fix join key expressions and add test coverage

This commit fixes a bug in the join-key expression grammar
and adds a test to cover it.  There was ambiguity that was not
properly tested in the previous commit.  The ambiguity is resolved
by using parenthesis to wrap an expression for the join key
compared to a direct field reference.  There may be a better
syntactic solution but it will need some creative thought and
we can live with this for now.