zq update through "fix and test for parallelized time grouping" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1650, authored by @alfred-landrum,
has been merged.

fix and test for parallelized time grouping

When we parallelize a query with a time grouping, ensure the input sort order direction is passed to the final merge operation. Also fixes a "sequential proc without procs" error that appeared if a bare reducer was passed into `zar zq`.

Fixes brimsec/zq#1641