zq update through "Fix ZNG spec grammar and TOC entry" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1653, authored by @philrz,
has been merged.

Fix ZNG spec grammar and TOC entry

A community user spotted an incomplete sentence in the coercion text at the bottom of the ZNG spec. Per a comment at the top of that appendix, I'm aware this text is likely to move elsewhere and that topic is tracked in brimsec/zq#1408. However, since we're actively changing some stuff in this area while refining expressions/etc., I figure we won't get to that for a bit. In the meantime, since people are noticing, here I'm fixing it where it current lives in the ZNG spec.

Fixes brimsec/zq#1647.