zq update through "zqd auto compaction" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1625, authored by @mattnibs,
has been merged.

zqd auto compaction

A simple first pass at providing auto-compaction functionality to zqd with
a simple hueristic: A compaction is run when an archivestore is first opened or
after data has been written to the archivestore. If new data is written to the
archive store while a compaction is running the current compaction task will be
canceled.

Add the task package- a simple interface for running background tasks in zqd.
The interface allows a space to initiate a SpaceOp for each background task and
provided the task with the op's context. This ensures that on zqd shutdown or
space deletion the background tasks will shutdown.

Closes brimsec/zq#1555