zq update through "Deterministic ordered merge" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1663, authored by @mattnibs,
has been merged.

Deterministic ordered merge

In an ordered merge if two records have the compare keys, use the results of
bytes.Compare to settle the tie and get deterministic results.

Also make the sort on archive.Import deterministic.

Closes brimsec/zq#1215