zq update through "unify the three variants of AST" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1665, authored by @mccanne,
has been merged.

unify the three variants of AST

This commit unifies the parsed ASTs so that whether an AST is
generated by pigeon, pegjs, or by marshaling an ast.Proc, the
resulting JSON is deep equal for the same zql query text.
Previously, different versions treated the zero and empty values
differently.

This update involves changes to the peg grammar actions
as well as the go ast package.  We also updated the ast command
so that it can display these three variations, as previously,
there was no way to display the pigeon version of the ast.

We updated the unit tests in zq/zql to compare the three variants
and added a test that looks through all the ztest yaml files to find
zqls to run and run the three-way comparison.

Fixes brimsec/zq#1638