zq update through "Updated perf numbers as of zq commit 1813bdf8" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1671, authored by @philrz,
has been merged.

Updated perf numbers as of zq commit 1813bdf8

These are a re-run of the perf numbers now that brimsec/zq#1669 has been addressed. The main difference to highlight is in the `jq` numbers. tl;dr on brimsec/zq#1669 - The `time` command was incorrectly only watching the `zcat` process before, and now that's been fixed to watch over the whole of the pipeline. Because of this, the `real` time reported for the `jq` operations has not changed significantly because `zcat` would hang until the whole pipeline had completed. But now that `jq`'s contribution is included in the overall, that explains the higher `user` and `sys` numbers. The same is true for the `zeek-cut` command lines as well, but `zeek-cut` is so dead simple that it's less noticeable.

On the first pass at this effort, the results revealed an unrelated perf regression in `zq` performance that was subsequently tracked in brimsec/zq#1670, fixed, and these numbers reflect the recovered performance after that fix.