Precise error message zqd pcap not found error (#1235)
If a pcap file is posted to a remote server the remote server will return 404
not found. The current error message was just name of the file. Update the
error message in this case to reflect that the server was not have to access
the file.

Related change:

- The zealot fetcher was not throwing an error if a stream request response was
a 404. Change so it does if the response is not ok.

Closes #1221