zq update through "fix bug in translating aliased types between type contexts" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1683, authored by @mccanne,
has been merged.

fix bug in translating aliased types between type contexts

When translating an alias from one type context to another,
you need to translate the aliased type to the target context
before looking up the alias name in the local context to establish
the binding.