zq update through "archive.Writer: fix staleDuration" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1696, authored by @mattnibs,
has been merged.

archive.Writer: fix staleDuration

Previously archive.Writer was setting the flusher interval to whatever the
flush timeout was. This didn't make sense and was also causing import test
flakiness. Set the flush interval to an arbitrary one second and make
staleDuration configurable.