zq update through "Say "REST" in place of "REST-like"" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1712, authored by @philrz,
has been merged.

Say "REST" in place of "REST-like"

In reviewing a draft blog post from @orochford I pointed out that describing our API as "REST-like" seemed strange, to which he responded that he was just mimicking what we said ourselves in these README files. @alfred-landrum chimed in to imply that we could go to just saying flat out "REST", so I'm proposing that change here.