zq update through "small refactor of groupby to get rid of params" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1720, authored by @mccanne,
has been merged.

small refactor of groupby to get rid of params

This is a small change to the groupby proc, which moves ast- and
compiler-specific logic to proc/compiler and eliminates groupby.Params.
This is a baby step toward the compiler refactor (issue brimsec/zq#1162)
and the bug fix for output type tracking (issue brimsec/zq#1701).