zq update through "allow len() function to work on IPs" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1725, authored by @mccanne,
has been merged.

allow len() function to work on IPs

This commit extends the len function to return the length of IP
addresses and networks.  This is handy for differentiating between
IPv4 (len=4) and IPv6 (len=16).  Likewise, networks are len 8 and
len 32.