zq update through "use flat keys, add comments, and improve names in groupby proc" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1728, authored by @mccanne,
has been merged.

use flat keys, add comments, and improve names in groupby proc

This commit takes a step torward the fix for brimsec/zq#1701 by rearranging
the group-by hash table keys to be stored as flattened values.
This will let the output stage build the output values based on
possibly changing output types.

We also added a comment on the spill-merge step and one relating
to memory pressure management of the hash table.

Finally, we adopted the names valRefs and keyRefs for the expr.Evaluators
that refer to the key/value fields themselves as compared to the expressions
that compute the values that go in those fields.