zq update through "eliminate extra copy of groupby keys" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1732, authored by @mccanne,
has been merged.

eliminate extra copy of groupby keys

Previously, the groupby code stored the key values in the Go map
as its string key and stored a second copy of the key values in the
groupby.Row value of the Go map.  This commit gets rid of the second
copy and uses the Go map hash key as the source for the key values.