zq update through "zson spec" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1715, authored by @mccanne,
has been merged.

zson spec

This commit adds a markdown description of the new ZSON format.
References to tzng have been removed from the ZNG spec and replaced
with references to the ZSON spec.  The language in the ZNG spec
has been further simplified a bit assuming knowledge of the ZSON
data model described in the new spec.

The zsonio writer has been improved to fix bugs and follow a few
changes that happend to the ZSON format.

We also made a first cut at updating the Zeek compat doc but we need
to have the ZSON reader before this can be finished.  We should also
consider explaining compat in terms of ZSON instead of ZNG.

Closes brimsec/zq#1662 brimsec/zq#1678