zq update through "add skip flag to ztest" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1746, authored by @mccanne,
has been merged.

add skip flag to ztest

This commit adds a yaml flag to ztest (skip: true) to tell ztest
to skip a particular test.  This is useful for test-driven dev or
when changes break tests and you want to disable certain tests
while you get them working again one at a time.