zq update through "Add default import stale duration" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1750, authored by @mattnibs,
has been merged.

Add default import stale duration

Currently if StaleDuration for an archive import was not getting set (as is the
case for the zar command and zqd post to archive store), the StaleDuration
would stay at 0. A StaleDuration of 0 means that pretty much all the data in
the archive would get flushed every 1 second (the interval at which the flusher
checks for new data to flush).

Set a default staleDuration to a more appropriate 5 seconds for an archive
Writer.