zq update through "Fix microindex trailer read bug" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1748, authored by @mattnibs,
has been merged.

Fix microindex trailer read bug

The read trailer part of microindex was relying on the call to the
io.Reader.Read to fill the entire buffer. For the aws sdk this often do not
happen.

According to the golang io.Reader.Read documentation:

If some data is available but not len(p) bytes, Read conventionally returns
what is available instead of waiting for more.

Use io.ReadFull to ensure the entire trailer is read.

Closes brimsec/zq#1730