zq update through "refactor space manager" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1693, authored by @alfred-landrum,
has been merged.

refactor space manager

This change moves & refactors the code responsible for space operations. Properties of spaces (name, storage location, etc) are now stored within a single file, which is treated similar to a database.

For archive based storage, we don't keep any state within zqd between interactions with the archive. For file based storage, as commented in apiserver.Manager, we do retain instances due to the non-persisted seek indexes we use to speed up queries for sub-portions of the space.

I've run the Brim tests against this code locally, btw, and verified they passed.