zq update through "Quote vars in Makefile that might include space chars when expanded" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1747, authored by @philrz,
has been merged.

Quote vars in Makefile that might include space chars when expanded

I had reason to run `make generate` for the first time ever. It failed with an error that took me a while to figure out. Turns out the root cause was because my `$PATH` happens to include an entry with a space character in it. Much like in shell scripting, it seems like quotes are in order.

Having seen & addressed that, I got to looking closer and realized the references to `CURDIR` would fail similarly if someone happened to have zq checked out in a location on the filesystem whose path had a space char in it. See for yourself:

```
14:06:37-phil@PhilMac:~/has space/zq$ make test-system
/bin/sh: space/zq/bin: No such file or directory
make: *** [bin/minio] Error 127
```

I've addressed those cases here as well.