zq update through "first small step toward compiler refactor" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1751, authored by @mccanne,
has been merged.

first small step toward compiler refactor

This commit is the first small move toward moving the compilation logic
and ast-dependent transformations in zq/compiler.  This commit moves
proc/compiler to the top-level and merges in the compilation and
optimization code from driver.  This is the first step on issue brimsec/zq#1162.