zq update through "move all ast references out of the procs and into the compiler package" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1754, authored by @mccanne,
has been merged.

move all ast references out of the procs and into the compiler package

This commit removes all remaining vestiges of "compilation" patterns
that operate on AST data structures out of the various proc packages
and into the compiler package.  This is step two of the compiler
refactoring issue brimsec/zq#1162.