zq update through "Reorder columns for Suricata alerts" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1749, authored by @philrz,
has been merged.

Reorder columns for Suricata alerts

In a recent group discussion it was noted that the current left-to-right order of the field columns for Suricata alerts is in some ways suboptimal. This PR proposes some changes.

Everyone seemed to agree the `event_type` field should move directly to the right of the `ts` field. This would make it fairly apples-to-apples with how we present Zeek events, since those also has `ts` as the leftmost column, followed by `_path` which is roughly equivalent to what `event_type` is for Suricata. This should make it easier for us to make a visually appealing colored tile/icon for alerts that will stand out nicely relative to Zeek events.

We also noted that some of the fields within the `alert` record are likely to be seen as more useful relative to others, such as `alert.severity`, `alert.category`, etc., so I've moved some of those leftward as well. Even though the `alert.metadata` sub-fields are empty much of the time, the way things work in zq today, they need to be kept adjacent to the contents of the `alert` record, so I've just kept those rightward. That means other fields like `app_proto`, `dest_ip`, etc. are quite far out to the right after the often-empty columns of `alert.metadata`. Thankfully the contents of those fields are quite redundant with the contents of the Zeek events we'll be linking to for context.

The only one I'm on the fence about is `community_id`. It was originally all the way rightward, which is where it happens to land today on Zeek `conn` events as well, so I was prepared to leave it there for symmetry. However, the Zeek events have their `uid` fields that are very leftward (directly to the right of `_path`) and `community_id` is the closest Suricata has to an equivalent concept, so I can see the case for bringing it leftward as I've done here (I think @orochford was in favor of this). If a user had other tools were cutting & pasting Community ID values into, having it prominently visible in a default view would be helpful. OTOH once we have the correlation in Log Detail that unites all Zeek events and Suricata alerts via the combo of `uid` and `community_id`, users may never have to _see_ the Community ID value to obtain its benefits, so that would be a reason to let it hide off to the right and keep the main screen real estate for more `alert` contents.

With the changes in this branch, here's an example of what a a full screen of alerts might look like with these leftward columns.

![image](https://user-images.githubusercontent.com/5934157/101097306-d0797a80-3575-11eb-8cb1-3c28dd8d3fc5.png)

Since that doesn't all fit, here's a top-down view. The left-to-right scrolling order would be the same as the top-down order shown here.

![image](https://user-images.githubusercontent.com/5934157/101097446-0454a000-3576-11eb-9176-b6db36f98321.png)