zq update through "improve zson types and add zson methods to zng.Type" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1765, authored by @mccanne,
has been merged.

improve zson types and add zson methods to zng.Type

This commit improves the type system for zson.  It updates the spec
and improves the implementation.  We also added ZSON() and ZSONOf()
methods to zng.Type.  Support for enums is now implemented.

The TypeType implementation has been changed to use a complex-type approach
(even the type values are serialized as primitive values in zng)
so that zng.Type pointers may be stored for each unique value in the context,
making the implementation simpler and more efficient.  We also
added a hash lookup to resolver.Context to map zng.Type pointers to
their string encoding so they don't have to be reformatted each time
typeof() is called on a given type.

Closes brimsec/zq#1675, brimsec/zq#1760