zq update through "Fix cutter type bug and refactor into cut, drop, and pick" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1773, authored by @mccanne,
has been merged.

Fix cutter type bug and refactor into cut, drop, and pick

This commit fixes the output type bug for cut that affected groupby
and cut	where the logic incorrectly assumed that the input record type
reliably determined the	output record type.  Since this doesn't
always hold, zng could be generated that corresponding to the wrong
type resulting in corrupt zng output.

While we refactored this, we realized the semantics of cut and "cut -c"
are really different and after discussion decided to adopt "drop"
as a replacement for "cut -c".	We also	debated	whether	cut semantics
should be like sort and	groupby	key expressions, which skip records
for which all input fields aren't present instead of cuts current
semantics, which allows "partial" cuts to occur when not all fields
are present.  We decided to keep the cut semantics and introduce
a new operator called "pick" that has the strict semantics.

Another	improvement here is that we created a generic "apply" proc
that allows any function to be implemented as a proc and the implementations
of cut and pick now use this pattern.

Fixes brimsec/zq#1739