zq update through "Use single quotes around the jq arg in perf script" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1787, authored by @philrz,
has been merged.

Use single quotes around the jq arg in perf script

The bottom two `jq` command lines in the Markdown table for the [performance summary](https://github.com/brimsec/zq/tree/master/performance) are wrong because of the overuse of double quotes. Here I've fixed the existing tables and also the script that generates the Markdown table so they'll come out correct the next time we regenerate these numbers.