zq update through "Add pick docs and cross-ref with cut" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1788, authored by @philrz,
has been merged.

Add pick docs and cross-ref with cut

This PR adds ZQL docs for the recently added `pick` processor. I've made sure to include an example that illustrates their fundamental difference and included links between the two so users who start at one but are seeking the stricter/looser behavior of the other can easily find it. I also swapped the order of two of the `cut` examples so the the numeric order matches to its comparable `pick` example.

I'm aware this introduces some redundant examples between `cut` and `pick`, but, so be it. I happen to prefer this approach to making a new user look in two places to get the full story. With what I'm doing here, I'm hoping if a user likes what they see in the first one they find, they're off to the races.